# Speed up sumIf/countIf