I hereby agree to the terms of the CLA available at: https://yandex.ru/legal/cla/?lang=en

Changelog category (leave one):
- Performance Improvement


Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Speed up sumIf and countIf aggregation functions

Detailed description / Documentation draft:

As https://github.com/ClickHouse/ClickHouse/pull/26248 showed, with certain types (integers) you can do some tricks to transform a flag check into a multiplication. Checking for null or checking for a different flag is exactly the same procedure (except null is 1 when you want to discard it, and -If is 1 when you want to keep it) so we can apply the same trick to sumIf.

Comparing the performance of a simple query:

```sql
SELECT sumIf(number, (number % 10) = 0)
FROM numbers(1000000000)
```

```
clickhouse-benchmark --max_threads 1 --query "Select sumIf(number, number % 10 == 0) from numbers(1000000000)"
```

* MASTER: localhost:9000, queries 22, QPS: 1.046, RPS: 1046078712.498, MiB/s: 7980.947, result RPS: 1.046, result MiB/s: 0.000.
* CHANGES: localhost:9000, queries 22, QPS: 1.139, RPS: 1138836403.326, MiB/s: 8688.632, result RPS: 1.139, result MiB/s: 0.000.
* IMPROVEMENT: **1.09x as fast**


Once you have this in place, you can use the same trick when you have a sumIf of a nullable type, except that this time you don't have 1 flag but 2, but combining them is trivial. 

```sql
SELECT sumIf(number, (number % 10) = 0)
FROM
(
    SELECT toNullable(number) AS number
    FROM numbers(1000000000)
)
```

```
clickhouse-benchmark --max_threads 1 --query "Select sumIf(number, number % 10 == 0) FROM (Select toNullable(number) as number from numbers(1000000000))"
```

* MASTER: localhost:9000, queries 11, QPS: 0.366, RPS: 366413296.245, MiB/s: 2795.512, result RPS: 0.366, result MiB/s: 0.000.
* CHANGES: localhost:9000, queries 12, QPS: 0.728, RPS: 727744569.722, MiB/s: 5552.250, result RPS: 0.728, result MiB/s: 0.000.
* IMPROVEMENT:  **1.99x as fast**


To be able to reach this point, it was necessary to specialize AggregateFunctionIfNullUnary::addBatchSinglePlace to call the nested function addBatchSinglePlaceNotNull with the appropiate data, so accidentally it also improves other functions like countIf:

```sql
SELECT countIf((number % 10) = 0)
FROM
(
    SELECT toNullable(number) AS number
    FROM numbers(1000000000)
)

```

```
clickhouse-benchmark --max_threads 1 --query "Select countIf(number % 10 == 0) FROM (Select toNullable(number) as number from numbers(1000000000))"
```

* MASTER: localhost:9000, queries 29, QPS: 0.355, RPS: 355401345.478, MiB/s: 2711.497, result RPS: 0.355, result MiB/s: 0.000.
* CHANGES: localhost:9000, queries 22, QPS: 1.106, RPS: 1105876088.463, MiB/s: 8437.165, result RPS: 1.106, result MiB/s: 0.000.
* IMPROVEMENT: **3.11x as fast**


I haven't looked at whether the performance tests are checking these functions (sumIf, sumIf with nullable, countIf with nullable) so I might need to add them or even break the PR into 2, one for each improvement. Let's see what the CI thinks about it first.