@Algunenano 2021-08-30
There was an issue in `countBytesInFilterWithNull` which provides the "simd" count+If+Nullable, but prior to this changes the function wasn't used so it wasn't broken.

Fixed the issue found in tests and added the flag merge in the If function too so it's applied in more functions automagically.

A great example of the improvement can be seen when calling sumIf(x, constant):
* `clickhouse-benchmark --max_threads 1 --query "SELECT sumIf(number, 1)　FROM　(　    SELECT toNullable(number) AS number　    FROM numbers(1000000000)　)"`
* Master: `localhost:9000, queries 11, QPS: 0.308, RPS: 308251206.100, MiB/s: 2351.770, result RPS: 0.308, result MiB/s: 0.000.`
* Changes:`localhost:9000, queries 43, QPS: 3.053, RPS: 3052875154.085, MiB/s: 23291.589, result RPS: 3.053, result MiB/s: 0.000.`
* Improvement: **9.92x as fast**